requirements.txt
flask
mysql-connector-python
scikit-learn
numpy
gunicorn
flask-socketio
flask-bcrypt
qrcode
eventlet
Pillow>=11.0.0
matplotlib
